NOISSUE - rename bashflux to cli and fix cert path
